style: #55: move out app config on upper level
